迁移 mip-map
