+++ components/mip-map/example/mip-map.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
   <title>mip-map MIP page</title>
   <link rel="canonical" href="对应的原页面地址">
-  <link rel="stylesheet" href="https://bos.nj.bpc.baidu.com/assets/mip/projects/mip.css">
+  <link rel="stylesheet" href="https://c.mipcdn.com/static/v2/mip.css">
   <style mip-custom>
     /* 自定义样式 */
 
@@ -18,8 +18,7 @@
     <script type="application/json">
       {
         "ak": "hKhuzfFBrcL6zGm4s6b371NDxaUrhFPl",
-        "hide-map": false,
-        "get-position": true,
+        "get-position": false,
         "location": {
           "province": "北京",
           "city": "北京市",
@@ -27,8 +26,13 @@
           "street": "故宫博物馆"
         },
         "controls": {
-          "NavigationControl": {},
-          "MapTypeControl": {}
+          "NavigationControl": {
+            "showZoomInfo": true,
+            "enableGeolocation": true
+          },
+          "MapTypeControl": {
+            "type": "BMAP_MAPTYPE_CONTROL_HORIZONTAL"
+          }
         },
         "info": {
           "width": 250,
@@ -36,10 +40,9 @@
           "content": "<h4>故宫博物馆</h4><p>地址：北京市东城区景山前街4号<br/>电话：(010)65131892</p></div>"
         }
       }
-
     </script>
   </mip-map>
-  <script src="https://bos.nj.bpc.baidu.com/assets/mip/projects/mip.js"></script>
+  <script src="https://c.mipcdn.com/static/v2/mip.js"></script>
   <script src="/mip-map/mip-map.js"></script>
 </body>
