+++ components/mip-map/mip-map.less
@@ -1,0 +1,10 @@
+mip-map {
+  #allmap {
+    width: 100%;
+    height: 100%;
+
+    img {
+      width: auto;
+    }
+  }
+}
